Shuffle the deckchairs to be more idiomatic.
