+++ pkg/cmd/auth/auth.go
@@ -1,4 +1,4 @@
-package main
+package auth
 
 import (
 	"bufio"
